REFACTOR: Move calendar view to views blueprint
- Created views blueprint for UI/visual components
- Migrated calendar route with complex task aggregation logic
- Maintained all calendar functionality including:
  - Date range calculations and calendar grid layout
  - Task filtering for project and independent tasks
  - Proper task data serialization for frontend
- Reduced app.py routes from 100 to 69 (31 routes migrated total)
- Systematic blueprint organization by functional area
